- Create a separate Form to allow designer layout of the messagebox - Simplified the MessageBox code - Fixed usage of icons to be system or Krypton - Use system logo
Fixes #242
