+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/KryptonMessageBoxForm.cs
@@ -1,0 +1,694 @@
+﻿#region BSD License
+/*
+ * 
+ * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
+ *  © Component Factory Pty Ltd, 2006 - 2016, (Version 4.5.0.0) All rights reserved.
+ * 
+ *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
+ *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2017 - 2021. All rights reserved. 
+ *
+ */
+#endregion
+
+
+namespace Krypton.Toolkit
+{
+    internal partial class KryptonMessageBoxForm : KryptonForm
+    {
+        #region Static Fields
+        private const int GAP = 10;
+        private static readonly int OS_MAJOR_VERSION;
+        #endregion
+        #region Instance Fields
+        private readonly string _text;
+        private readonly string _caption;
+        private readonly MessageBoxButtons _buttons;
+        private readonly KryptonMessageBoxIcon _kryptonMessageBoxIcon;
+        private readonly MessageBoxIcon _messageBoxIcon;
+
+        private readonly MessageBoxDefaultButton _defaultButton;
+        private MessageBoxOptions _options; // TODO: What is this used for ? e.g. MessageBoxOptions.RTL
+        // If help information provided or we are not a service/default desktop application then grab an owner for showing the message box
+        private readonly IWin32Window _showOwner;
+        private readonly HelpInfo _helpInfo;
+
+        #endregion
+
+        #region Identity
+        static KryptonMessageBoxForm()
+        {
+            OS_MAJOR_VERSION = Environment.OSVersion.Version.Major;
+        }
+        public KryptonMessageBoxForm()
+        {
+            InitializeComponent();
+        }
+
+
+        internal KryptonMessageBoxForm(IWin32Window showOwner, string text, string caption,
+            MessageBoxButtons buttons, KryptonMessageBoxIcon icon,
+            MessageBoxDefaultButton defaultButton, MessageBoxOptions options,
+            HelpInfo helpInfo, bool? showCtrlCopy)
+        {
+            // Store incoming values
+            _text = text;
+            _caption = caption;
+            _buttons = buttons;
+            _kryptonMessageBoxIcon = icon;
+            _defaultButton = defaultButton;
+            _options = options;
+            _helpInfo = helpInfo;
+            _showOwner = showOwner;
+
+            // Create the form contents
+            InitializeComponent();
+
+            // Update contents to match requirements
+            UpdateText();
+            UpdateIcon();
+            UpdateButtons();
+            UpdateDefault();
+            UpdateHelp();
+            UpdateTextExtra(showCtrlCopy);
+
+            // Finally calculate and set form sizing
+            UpdateSizing(showOwner);
+        }
+
+        internal KryptonMessageBoxForm(IWin32Window showOwner, string text, string caption,
+            MessageBoxButtons buttons, MessageBoxIcon icon,
+            MessageBoxDefaultButton defaultButton, MessageBoxOptions options,
+            HelpInfo helpInfo, bool? showCtrlCopy)
+        {
+            // Store incoming values
+            _text = text;
+            _caption = caption;
+            _buttons = buttons;
+            _messageBoxIcon = icon;
+            _defaultButton = defaultButton;
+            _options = options;
+            _helpInfo = helpInfo;
+            _showOwner = showOwner;
+
+            // Create the form contents
+            InitializeComponent();
+
+            // Update contents to match requirements
+            UpdateText();
+            UpdateIcon();
+            UpdateButtons();
+            UpdateDefault();
+            UpdateHelp();
+            UpdateTextExtra(showCtrlCopy);
+
+            // Finally calculate and set form sizing
+            UpdateSizing(showOwner);
+
+        }
+        #endregion Identity
+
+        private void UpdateText()
+        {
+            Text = (string.IsNullOrEmpty(_caption) ? string.Empty : _caption.Split(Environment.NewLine.ToCharArray())[0]);
+            _messageText.Text = _text;
+        }
+
+        private void UpdateTextExtra(bool? showCtrlCopy)
+        {
+            if (!showCtrlCopy.HasValue)
+            {
+                switch (_kryptonMessageBoxIcon)
+                {
+                    case KryptonMessageBoxIcon.ERROR:
+                    case KryptonMessageBoxIcon.EXCLAMATION:
+                        showCtrlCopy = true;
+                        break;
+                }
+
+                switch (_messageBoxIcon)
+                {
+                    case MessageBoxIcon.Error:
+                    case MessageBoxIcon.Exclamation:
+                        showCtrlCopy = true;
+                        break;
+                }
+            }
+
+            if (showCtrlCopy == true)
+            {
+                TextExtra = @"Ctrl+c to copy";
+            }
+        }
+
+        private void UpdateIcon()
+        {
+            switch (_kryptonMessageBoxIcon)
+            {
+                default:
+                case KryptonMessageBoxIcon.NONE:
+                    // Windows XP and before will Beep, Vista and above do not!
+                    if (OS_MAJOR_VERSION < 6)
+                    {
+                        SystemSounds.Beep.Play();
+                    }
+
+                    break;
+                case KryptonMessageBoxIcon.QUESTION:
+                    _messageIcon.Image = MessageBoxResources.Question;
+                    SystemSounds.Question.Play();
+                    break;
+                case KryptonMessageBoxIcon.EXCLAMATION:
+                case KryptonMessageBoxIcon.INFORMATION:
+                    _messageIcon.Image = MessageBoxResources.MBInformation;
+                    SystemSounds.Asterisk.Play();
+                    break;
+                case KryptonMessageBoxIcon.WARNING:
+                    _messageIcon.Image = MessageBoxResources.Warning;
+                    SystemSounds.Exclamation.Play();
+                    break;
+                case KryptonMessageBoxIcon.ERROR:
+                    _messageIcon.Image = MessageBoxResources.Critical;
+                    SystemSounds.Hand.Play();
+                    break;
+                case KryptonMessageBoxIcon.ASTERISK:
+                    _messageIcon.Image = MessageBoxResources.Asterisk;
+                    SystemSounds.Asterisk.Play();
+                    break;
+                case KryptonMessageBoxIcon.HAND:
+                    _messageIcon.Image = MessageBoxResources.Hand;
+                    SystemSounds.Hand.Play();
+                    break;
+                case KryptonMessageBoxIcon.STOP:
+                    _messageIcon.Image = MessageBoxResources.Stop;
+                    SystemSounds.Hand.Play();
+                    break;
+                case KryptonMessageBoxIcon.SHIELD:
+                    _messageIcon.Image = SystemIcons.Shield.ToBitmap();
+                    break;
+                case KryptonMessageBoxIcon.WINDOWSLOGO:
+                    _messageIcon.Image = SystemIcons.WinLogo.ToBitmap();
+                    break;
+            }
+
+            switch (_messageBoxIcon)
+            {
+                default:
+                case MessageBoxIcon.None:
+                    // Windows XP and before will Beep, Vista and above do not!
+                    if (OS_MAJOR_VERSION < 6)
+                    {
+                        SystemSounds.Beep.Play();
+                    }
+                    break;
+                case MessageBoxIcon.Question:
+                    _messageIcon.Image = SystemIcons.Question.ToBitmap();
+                    SystemSounds.Question.Play();
+                    break;
+                case MessageBoxIcon.Error:
+                //case MessageBoxIcon.Hand:
+                //case MessageBoxIcon.Stop:
+                    _messageIcon.Image = SystemIcons.Error.ToBitmap();
+                    SystemSounds.Hand.Play();
+                    break;
+                case MessageBoxIcon.Warning: 
+                    //case MessageBoxIcon.Exclamation:
+                    _messageIcon.Image = SystemIcons.Warning.ToBitmap();
+                    SystemSounds.Exclamation.Play();
+                    break;
+                case MessageBoxIcon.Information:
+                // case MessageBoxIcon.Asterisk:
+                    _messageIcon.Image = SystemIcons.Information.ToBitmap();
+                    SystemSounds.Asterisk.Play();
+                    break;
+            }
+
+            _messageIcon.Visible = (_kryptonMessageBoxIcon != KryptonMessageBoxIcon.NONE) || (_messageBoxIcon != MessageBoxIcon.None);
+
+        }
+
+        private void UpdateButtons()
+        {
+            switch (_buttons)
+            {
+                case MessageBoxButtons.OK:
+                    _button1.Text = KryptonManager.Strings.OK;
+                    _button1.DialogResult = DialogResult.OK;
+                    _button1.Visible = true;
+                    _button1.Enabled = true;
+                    break;
+                case MessageBoxButtons.OKCancel:
+                    _button1.Text = KryptonManager.Strings.OK;
+                    _button2.Text = KryptonManager.Strings.Cancel;
+                    _button1.DialogResult = DialogResult.OK;
+                    _button2.DialogResult = DialogResult.Cancel;
+                    _button1.Visible = true;
+                    _button1.Enabled = true;
+                    _button2.Visible = true;
+                    _button2.Enabled = true;
+                    break;
+                case MessageBoxButtons.YesNo:
+                    _button1.Text = KryptonManager.Strings.Yes;
+                    _button2.Text = KryptonManager.Strings.No;
+                    _button1.DialogResult = DialogResult.Yes;
+                    _button2.DialogResult = DialogResult.No;
+                    _button1.Visible = true;
+                    _button1.Enabled = true;
+                    _button2.Visible = true;
+                    _button2.Enabled = true;
+                    ControlBox = false;
+                    break;
+                case MessageBoxButtons.YesNoCancel:
+                    _button1.Text = KryptonManager.Strings.Yes;
+                    _button2.Text = KryptonManager.Strings.No;
+                    _button3.Text = KryptonManager.Strings.Cancel;
+                    _button1.DialogResult = DialogResult.Yes;
+                    _button2.DialogResult = DialogResult.No;
+                    _button3.DialogResult = DialogResult.Cancel;
+                    _button1.Visible = true;
+                    _button1.Enabled = true;
+                    _button2.Visible = true;
+                    _button2.Enabled = true;
+                    _button3.Visible = true;
+                    _button3.Enabled = true;
+                    break;
+                case MessageBoxButtons.RetryCancel:
+                    _button1.Text = KryptonManager.Strings.Retry;
+                    _button2.Text = KryptonManager.Strings.Cancel;
+                    _button1.DialogResult = DialogResult.Retry;
+                    _button2.DialogResult = DialogResult.Cancel;
+                    _button1.Visible = true;
+                    _button1.Enabled = true;
+                    _button2.Visible = true;
+                    _button2.Enabled = true;
+                    break;
+                case MessageBoxButtons.AbortRetryIgnore:
+                    _button1.Text = KryptonManager.Strings.Abort;
+                    _button2.Text = KryptonManager.Strings.Retry;
+                    _button3.Text = KryptonManager.Strings.Ignore;
+                    _button1.DialogResult = DialogResult.Abort;
+                    _button2.DialogResult = DialogResult.Retry;
+                    _button3.DialogResult = DialogResult.Ignore;
+                    _button1.Visible = true;
+                    _button1.Enabled = true;
+                    _button2.Visible = true;
+                    _button2.Enabled = true;
+                    _button3.Visible = true;
+                    _button3.Enabled = true;
+                    ControlBox = false;
+                    break;
+            }
+
+            // Do we ignore the Alt+F4 on the buttons?
+            if (!ControlBox)
+            {
+                _button1.IgnoreAltF4 = true;
+                _button2.IgnoreAltF4 = true;
+                _button3.IgnoreAltF4 = true;
+                _button4.IgnoreAltF4 = true;
+            }
+        }
+
+        private void UpdateDefault()
+        {
+            switch (_defaultButton)
+            {
+                case MessageBoxDefaultButton.Button2:
+                    _button2.Select();
+                    break;
+                case MessageBoxDefaultButton.Button3:
+                    _button3.Select();
+                    break;
+            }
+        }
+
+        private void UpdateHelp()
+        {
+            if (_helpInfo == null)
+            {
+                return;
+            }
+
+            MessageButton helpButton;
+            switch (_buttons)
+            {
+                case MessageBoxButtons.OK:
+                    helpButton = _button2;
+                    break;
+                case MessageBoxButtons.OKCancel:
+                case MessageBoxButtons.YesNo:
+                case MessageBoxButtons.RetryCancel:
+                    helpButton = _button3;
+                    break;
+                case MessageBoxButtons.AbortRetryIgnore:
+                case MessageBoxButtons.YesNoCancel:
+                    helpButton = _button4;
+                    break;
+                default:
+                    throw new ArgumentOutOfRangeException();
+            }
+
+            if (helpButton != null)
+            {
+                helpButton.Visible = true;
+                helpButton.Enabled = true;
+                helpButton.Text = KryptonManager.Strings.Help;
+                helpButton.KeyPress += (_, _) => LaunchHelp();
+                helpButton.Click += (_, _) => LaunchHelp();
+            }
+        }
+
+        /// <summary>
+        /// When the user clicks the Help button, the Help file specified in the helpFilePath parameter
+        /// is opened and the Help keyword topic identified by the keyword parameter is displayed.
+        /// The form that owns the message box (or the active form) also receives the HelpRequested event.
+        /// </summary>
+        private void LaunchHelp()
+        {
+            try
+            {
+                Control control = FromHandle(_showOwner.Handle);
+
+                MethodInfo mInfoMethod = control.GetType().GetMethod(@"OnHelpRequested", BindingFlags.Instance | BindingFlags.NonPublic,
+                    Type.DefaultBinder, new[] { typeof(HelpEventArgs) }, null);
+                if (mInfoMethod != null)
+                {
+                    mInfoMethod.Invoke(control, new object[] { new HelpEventArgs(MousePosition) });
+                }
+                if (MissingFrameWorkAPIs.IsNullOrWhiteSpace(_helpInfo.HelpFilePath))
+                {
+                    return;
+                }
+
+                if (!MissingFrameWorkAPIs.IsNullOrWhiteSpace(_helpInfo.Keyword))
+                {
+                    Help.ShowHelp(control, _helpInfo.HelpFilePath, _helpInfo.Keyword);
+                }
+                else
+                {
+                    Help.ShowHelp(control, _helpInfo.HelpFilePath, _helpInfo.Navigator, _helpInfo.Param);
+                }
+            }
+            catch
+            {
+                // Do nothing if failure to send to Parent
+            }
+
+        }
+
+        private void UpdateSizing(IWin32Window showOwner)
+        {
+            Size messageSizing = UpdateMessageSizing(showOwner);
+            Size buttonsSizing = UpdateButtonsSizing();
+
+            // Size of window is calculated from the client area
+            ClientSize = new Size(Math.Max(messageSizing.Width, buttonsSizing.Width),
+                                  messageSizing.Height + buttonsSizing.Height);
+        }
+
+        private Size UpdateMessageSizing(IWin32Window showOwner)
+        {
+            // Update size of the message label but with a maximum width
+            Size textSize;
+            using (Graphics g = CreateGraphics())
+            {
+                // Find size of the label, with a max of 2/3 screen width
+                Screen screen = showOwner != null ? Screen.FromHandle(showOwner.Handle) : Screen.PrimaryScreen;
+                SizeF scaledMonitorSize = screen.Bounds.Size;
+                scaledMonitorSize.Width *= (2 / 3.0f);
+                scaledMonitorSize.Height *= 0.95f;
+                _messageText.UpdateFont();
+                SizeF messageSize = g.MeasureString(_text, _messageText.Font, scaledMonitorSize);
+                // SKC: Don't forget to add the TextExtra into the calculation
+                SizeF captionSize = g.MeasureString($@"{_caption} {TextExtra}", _messageText.Font, scaledMonitorSize);
+
+                float messageXSize = Math.Max(messageSize.Width, captionSize.Width);
+                // Work out DPI adjustment factor
+                float factorX = g.DpiX > 96 ? ((1.0f * g.DpiX) / 96) : 1.0f;
+                float factorY = g.DpiY > 96 ? ((1.0f * g.DpiY) / 96) : 1.0f;
+                messageSize.Width = messageXSize * factorX;
+                messageSize.Height = messageSize.Height * factorY;
+
+                // Always add on ad extra 5 pixels as sometimes the measure size does not draw the last 
+                // character it contains, this ensures there is always definitely enough space for it all
+                messageSize.Width += 5;
+                textSize = Size.Ceiling(messageSize);
+            }
+
+            // Find size of icon area plus the text area added together
+            if (_messageIcon.Image != null)
+            {
+                //_messageIcon.Width = _messageIcon.Image.Width;
+                //_messageIcon.Height = Math.Max(_messageIcon.Image.Height, textSize.Height);
+                return new Size(textSize.Width +_messageIcon.Width, Math.Max(_messageIcon.Height + 10, textSize.Height));
+            }
+
+            return textSize;
+        }
+
+        private Size UpdateButtonsSizing()
+        {
+            int numButtons = 1;
+
+            // Button1 is always visible
+            Size button1Size = _button1.GetPreferredSize(Size.Empty);
+            Size maxButtonSize = new(button1Size.Width + GAP, button1Size.Height);
+
+            // If Button2 is visible
+            if (_button2.Enabled)
+            {
+                numButtons++;
+                Size button2Size = _button2.GetPreferredSize(Size.Empty);
+                maxButtonSize.Width = Math.Max(maxButtonSize.Width, button2Size.Width + GAP);
+                maxButtonSize.Height = Math.Max(maxButtonSize.Height, button2Size.Height);
+            }
+
+            // If Button3 is visible
+            if (_button3.Enabled)
+            {
+                numButtons++;
+                Size button3Size = _button3.GetPreferredSize(Size.Empty);
+                maxButtonSize.Width = Math.Max(maxButtonSize.Width, button3Size.Width + GAP);
+                maxButtonSize.Height = Math.Max(maxButtonSize.Height, button3Size.Height);
+            }
+            // If Button4 is visible
+            if (_button4.Enabled)
+            {
+                numButtons++;
+                Size button4Size = _button4.GetPreferredSize(Size.Empty);
+                maxButtonSize.Width = Math.Max(maxButtonSize.Width, button4Size.Width + GAP);
+                maxButtonSize.Height = Math.Max(maxButtonSize.Height, button4Size.Height);
+            }
+
+            // Start positioning buttons 10 pixels from right edge
+            int right = _panelButtons.Right - GAP;
+
+            // If Button4 is visible
+            if (_button4.Enabled)
+            {
+                _button4.Location = new Point(right - maxButtonSize.Width, GAP);
+                _button4.Size = maxButtonSize;
+                right -= maxButtonSize.Width + GAP;
+            }
+
+            // If Button3 is visible
+            if (_button3.Enabled)
+            {
+                _button3.Location = new Point(right - maxButtonSize.Width, GAP);
+                _button3.Size = maxButtonSize;
+                right -= maxButtonSize.Width + GAP;
+            }
+
+            // If Button2 is visible
+            if (_button2.Enabled)
+            {
+                _button2.Location = new Point(right - maxButtonSize.Width, GAP);
+                _button2.Size = maxButtonSize;
+                right -= maxButtonSize.Width + GAP;
+            }
+
+            // Button1 is always visible
+            _button1.Location = new Point(right - maxButtonSize.Width, GAP);
+            _button1.Size = maxButtonSize;
+
+            // Size the panel for the buttons
+            _panelButtons.Size = new Size((maxButtonSize.Width * numButtons) + (GAP * (numButtons + 1)), maxButtonSize.Height + (GAP * 2));
+
+            // Button area is the number of buttons with gaps between them and 10 pixels around all edges
+            return new Size((maxButtonSize.Width * numButtons) + (GAP * (numButtons + 1)), maxButtonSize.Height + (GAP * 2));
+        }
+
+        private void button_keyDown(object sender, KeyEventArgs e)
+        {
+            // Escape key kills the dialog if we allow it to be closed
+            if ((e.KeyCode == Keys.Escape) && ControlBox)
+            {
+                Close();
+            }
+            else
+            {
+                // Pressing Ctrl+C should copy message text into the clipboard
+                if ((e.Modifiers == Keys.Control) && (e.KeyCode == Keys.C))
+                {
+                    StringBuilder sb = new();
+
+                    sb.AppendLine("---------------------------");
+                    sb.AppendLine(Text);
+                    sb.AppendLine("---------------------------");
+                    sb.AppendLine(_messageText.Text);
+                    sb.AppendLine("---------------------------");
+                    sb.Append(_button1.Text);
+                    sb.Append("   ");
+                    if (_button2.Enabled)
+                    {
+                        sb.Append(_button2.Text);
+                        sb.Append("   ");
+                        if (_button3.Enabled)
+                        {
+                            sb.Append(_button3.Text);
+                            sb.Append("   ");
+                        }
+                        if (_button4.Enabled)
+                        {
+                            sb.Append(_button4.Text);
+                            sb.Append("   ");
+                        }
+                    }
+                    sb.AppendLine("");
+                    sb.AppendLine("---------------------------");
+
+                    Clipboard.SetText(sb.ToString(), TextDataFormat.Text);
+                    Clipboard.SetText(sb.ToString(), TextDataFormat.UnicodeText);
+                }
+            }
+        }
+
+    }
+    #region Types
+    internal class HelpInfo
+    {
+        #region Instance Fields
+
+        #endregion
+
+        #region Identity
+
+        /// <summary>
+        /// Initialize a new instance of the HelpInfo class.
+        /// </summary>
+        /// <param name="helpFilePath">Value for HelpFilePath.</param>
+        /// <param name="keyword">Value for Keyword</param>
+        public HelpInfo(string helpFilePath = null, string keyword = null)
+        : this(helpFilePath, keyword, !MissingFrameWorkAPIs.IsNullOrWhiteSpace(keyword) ? HelpNavigator.Topic : HelpNavigator.TableOfContents, null)
+        {
+
+        }
+
+        /// <summary>
+        /// Initialize a new instance of the HelpInfo class.
+        /// </summary>
+        /// <param name="helpFilePath">Value for HelpFilePath.</param>
+        /// <param name="navigator">Value for Navigator</param>
+        /// <param name="param"></param>
+        public HelpInfo(string helpFilePath, HelpNavigator navigator, object param = null)
+            : this(helpFilePath, null, navigator, param)
+        {
+
+        }
+
+        /// <summary>
+        /// Initialize a new instance of the HelpInfo class.
+        /// </summary>
+        /// <param name="helpFilePath">Value for HelpFilePath.</param>
+        /// <param name="navigator">Value for Navigator</param>
+        /// <param name="keyword">Value for Keyword</param>
+        /// <param name="param"></param>
+        private HelpInfo(string helpFilePath, string keyword, HelpNavigator navigator, object param)
+        {
+            HelpFilePath = helpFilePath;
+            Keyword = keyword;
+            Navigator = navigator;
+            Param = param;
+        }
+        #endregion
+
+        #region Properties
+        /// <summary>
+        /// Gets the HelpFilePath property.
+        /// </summary>
+        public string HelpFilePath { get; }
+
+        /// <summary>
+        /// Gets the Keyword property.
+        /// </summary>
+        public string Keyword { get; }
+
+        /// <summary>
+        /// Gets the Navigator property.
+        /// </summary>
+        public HelpNavigator Navigator { get; }
+
+        /// <summary>
+        /// Gets the Param property.
+        /// </summary>
+        public object Param { get; }
+
+        #endregion
+    }
+
+    #endregion
+
+    [ToolboxItem(false)]
+    [DesignTimeVisible(false)]
+    internal class MessageButton : KryptonButton
+    {
+
+        #region Instance Fields
+
+        #endregion
+
+        #region Identity
+        public MessageButton()
+        {
+            IgnoreAltF4 = false;
+            Visible = false;
+            Enabled = false;
+        }
+
+        /// <summary>
+        /// Gets and sets the ignoring of Alt+F4
+        /// </summary>
+        public bool IgnoreAltF4 { get; set; }
+
+        #endregion
+
+        #region Protected
+        /// <summary>
+        /// Processes Windows messages.
+        /// </summary>
+        /// <param name="m">The Windows Message to process. </param>
+        protected override void WndProc(ref Message m)
+        {
+            switch (m.Msg)
+            {
+                case PI.WM_.KEYDOWN:
+                case PI.WM_.SYSKEYDOWN:
+                    if (IgnoreAltF4)
+                    {
+                        // Extract the keys being pressed
+                        Keys keys = ((Keys)((int)m.WParam.ToInt64()));
+
+                        // If the user standard combination ALT + F4
+                        if ((keys == Keys.F4) && ((ModifierKeys & Keys.Alt) == Keys.Alt))
+                        {
+                            // Eat the message, so standard window proc does not close the window
+                            return;
+                        }
+                    }
+                    break;
+            }
+
+            base.WndProc(ref m);
+        }
+        #endregion
+    }
+}
